requirements.txt
Pillow==10.0.0
pyperclip3~=0.4.1
